Disable deferred updates on the stream and standalone annotation pages (#112)
The deferred update UI was added primarily to avoid the problem of
annotation updates being distracting while a user is composing
annotations in the sidebar. This is less relevant on the stream.

Additionally fix a bug where updates from unfocused groups were not
shown on the stream, since the stream shows updates from all groups.