Count pending deletions as pending updates (#114)
When calculating the count of pending updates that is used to determine
whether to show the 'Apply Updates' icon or not, count deletions.

In addition to changing the count calculation itself, this also requires
that pending deletions are discarded for annotations that are not
currently loaded. Such notifications may be received because the
deletion may have happened in an un-focused group. Unlike create/update
notifications, no group information is included in the notification
itself so we look up the ID in the local store.