Do not try to setup communication to host frame on stream
This fixes an error when trying to load the Hypothesis client on the
stream or standalone annotation pages, caused by both the stream app and
the sidebar app trying to set up a 'Discovery' server for annotation
'guests' to connect to.

With this change, only the sidebar app will set up a Discovery server
and hence it is now possible to annotate the stream and standalone
annotation pages.