Fix formatting changes made via toolbar buttons not persisting
When making formatting changes to the text of an annotation with the
toolbar button without subsequently making a manual edit before clicking
Preview or Save, the onEditText() callback was never triggered and so
the unsaved changes were reverted when exiting edit mode.

This happened because onEditText() was only fired in response to the
'input' event which only occurs in response to programmatic, not manual
changes to the input's value.

Fixes #147