Add a set of HTML anchoring tests with expected failures
Add a set of tests which check that Ranges within a DOM tree are
correctly 'described' as selectors and then successfully converted back to
Range objects by anchoring.

The tests are derived from the suite of test cases collected for the
Annotator project at
https://github.com/openannotation/annotator/blob/v1.2.x/test/spec/range_spec.coffee

These tests found failures in describing ranges with position and quote
selectors in many cases and with a range selector in one case. The test
lists and checks these expected failures.