Add PDF anchoring unit tests
Add a set of tests using a fake implementation of PDF.js'
PDFViewerApplication interface.

The initial set of tests simulate a nice born-digital PDF in which the
text layer of the PDF contains one <div> per line. In future we can
easily extend this to emulate the kind of structure that would be seen
when working with eg. an OCR-ed PDF