Add a new entry point script for the client application
Add a new bundle (build/boot.js) which serves as the main entry point
for the client. This entry point will be used in two places:

 1. When embedded in a host web page it will add the bundle for the
    annotation layer (injector.bundle.js) to the page plus supporting
    assets.

 2. When included as the sole asset in the sidebar's HTML page
    it will add the bundle for the sidebar app (app.bundle.js) to the
    page plus supporting assets.

Having both entry points be a single script means that when it is
referenced in app.html, the browser will already have it cached locally,
saving a roundtrip.

In this commit the script just adds all <script> and <link> tags for the
annotation layer or sidebar to the document. In future it can be made
smarter - eg. by preloading sidebar assets when loaded in the host page
and not loading polyfills in newer browsers.