Specify boot script as the entry point for the client
Add a "main" entry to package.json so that the boot script is treated as
the main URL for the package when loading it via services like
unpkg.com.

We are already using the "browser" field to specify how to resolve
modules for Browserify, so it cannot be set to the entry point directly,
although hacks [1] are possible.

[1] https://gist.github.com/defunctzombie/4339901#gistcomment-1582456