Modify live reload server to use client served by package content server
The live reload server's example page was modified to use a script tag
for config settings because the boot script currently only reads that
config source and does not use `window.hypothesisConfig()`.

In order to use the sidebar app from the local service, the service
needs to be configured to serve the local client by setting the
'CLIENT_URL' env var before running the devserver and in the client, the
'H_SERVICE_URL' env var needs to be pointed at the devserver host.