Make boot script load settings from correct config tags
An inconsistency crept in where the annotator code reads settings from
JSON script tags with a 'js-hypothesis-config' class name but the sidebar app
reads settings from JSON script tags with a 'js-hypothesis-settings'
class name. The boot script ended up using 'js-hypothesis-settings'
which is the default class name specified in `settings.js`.

We should use the same name everywhere, which needs to be
'js-hypothesis-config' since that is used by the public documentation
[1].

This commit changes the default class to `js-hypothesis-config` and
removes the ability to specify alternative class names when calling
`settings()`.

The extension and service will need to be updated to use the new class
name.

[1] See docs/config.md