Support onLogin callbacks from partner sites
When the client is embedded in a partner site and a login button in the
client is clicked, call the JavaScript onLogin function that was
provided to us by the partner via the window.hypothesisConfig function.

When not embedded in a partner site, just prompt for a first-party login
as normal.

Note that this means that if the client is embedded in a partner site
with no onLogin callback in window.hypothesisConfig, then nothing will
happen when the login button in the client is clicked.