Add `TextQuoteAnchor.toPositionAnchor` method
This method used to exist in the `dom-anchor-text-quote` library but got
lost when the class was moved to a wrapper in `types.coffee` after the
upgrade to dom-anchor-text-quote v3.x.

This method is used by PDF anchoring and its absence caused quote
anchoring to throw an exception. The majority of PDF annotations still
continued to anchor because the position anchor still worked.

This commit also adds some basic documentation to the `types` module and
some basic API tests.