Fix anchoring using quote selector alone failing
Looking at the `anchor` function, it is clearly intended that the
position selector is optional for anchoring, as is the case for HTML
anchoring. However, anchoring using only a quote selector did not
actually work. This was in part due to the bug fixed in the previous
commit but also because of a missing check for a null position selector
when populating the (quote,position) cache.

In order for the test to pass using only a quote, it was necessary to
change the search string to a unique quote which only occurs once in the
document. Otherwise the quote anchoring found the occurrence of
'Netherfield Park' on the non-rendered page 1.