Remove assumption that client asset URL contains '/build/' path segment
Remove the assumption that client asset URLs always take the form:

  ${ROOT_URL}/build/${PATH_FROM_MANIFEST}

In the Chrome extension, this is not the case.

Instead change the definition of the `assetRoot` config setting to be
the location where the contents of the `build/` dir are hosted, instead
of the root of the package itself.