Rename the onLogin callback to onLoginRequest
The trigger for invoking this callback function is the user requesting
that a login flow be started. In other words, at the point this callback
is invoked, a login has not occurred, and indeed may never occur if the
user does not complete the login flow.

As such, it seems clearer to have this callback called "onLoginRequest"
rather than "onLogin."

Fixes #274.