Permit queries via Url.
This project enables the user to include a default filter query as part of a url.

The syntax works with the #annotations: fragment.  If a Q: or Query: (case insensitive)
is included, whatever follows the query will be included in a search filter as a
default.

user tags may be included in this query string. For instance user:USERNAME will
filter for the user called USERNAME. Using an @hypothes.is will search for exactly
that user. So far, the user, tag or any tags are the main permissible items.

So http://www.example.com/path/to/file#annotations:Q:user:USERNAMEtag:TAGany:KEYWORD

will filter the page for annotations by USERNAME tagged TAG and containing KEYWORD.