Implement moderation banner for annotation cards (#285)
* Implement <moderation-banner> component

Implement the UI component for the moderation banner at the top of
annotation cards, following the designs on [1].

[1] https://github.com/hypothesis/product-backlog/issues/183

* Add reducer and actions for managing moderation app state

Add a module which defines the state and actions related to moderation
status of annotations.

* Display moderation banner at the top of annotation cards

The moderation banner is currently only supported on top-level
annotations. It will need to be adjusted and moved to a different
component in order to support replies.

* Support displaying moderation banner above replies.

 * Move the `<moderation-banner>` component to `<annotation-thread>` so
   that it can be displayed for replies as well as annotations.

 * Customize the styling of the banner when the annotation is a reply.

* Add a simple test for display change when moderation banner is used on a reply

* Add additional parameter documentation

* Convert `<moderation-banner>` to a component

As per hypothesis/client#292, all element directives are being converted
to components.

* Add tooltips to moderation banner buttons.

The "Hide" and "Undo" labels for the button were potentially ambiguous.
The final labels are still being discussed but improve the UI by adding
tooltips and changing "Undo" to "Unhide".

* Correct a reference to a button label in tests

The label changed from "Undo" to "Unhide".