Split "viewer" template into different templates for each app type
In preparation for converting the controllers which provide the logic
for the content area of the sidebar (below the top bar) into components,
split up the viewer.html template into one template for each application
type (sidebar, stream, single annotation page).

This allows removing a number of conditionals from the components in
each template.