Update moderation flag count after moderator flags an annotation
Flagging or unflagging an annotation previously updated the status
flag indicating whether an annotation had been flagged by the current
user but did not update the flag count if the user was also a moderator.

This needs to be done to make the moderation banner's state update when
a user who is also a moderator flags an annotation.