Use hyphens instead of underscores in template filenames
The use of underscores dates back to when the templates were part of the
service code and used the same naming conventions as the service's
Jinja2 templates.

Using hyphens makes the template filenames match the other files
which make up a component (JS & SCSS files). This makes it easier to
switch between the different parts when using a fuzzy file finder.