Show error on fail to get OAuth access token
Show an error if the initial request for an OAuth access token fails.

When the client is embedded in a page that includes a services
configuration setting (see <http://h.readthedocs.io/projects/client/en/latest/publishers/config/#cmdoption-arg-services>)
then it tries to login to Hypothesis using OAuth, exchanging the grant
token from the services config setting for an access token from the
Hypothesis API.

If this request for an access token fails (for example, if the
Hypothesis server responds with an error) then the client ends up in an
invalid state. Some of the sidebar components are not fully rendered,
with visible parts missing, no annotations are shown, and actions such
as trying to create an annotation fail in broken ways.

We should really fix the completely broken appearance and behavior of
the client in this state, but that'll be a big job.

For now, at least show an error message to the user telling them that
they need to reload the page.

There's currently no facility to retry the access token request but if
the user reloads the page the whole process will begin again and the
request will be tried again.