Remove the annotator.js vendor dependency
In order to avoid heavy refactoring some classes from Annotator were moved over to this codebase.
This includes the class for the `Document` plugin and Annotator’s base class `Delegator`.

The class `Guest` is now responsible for loading the remaining Annotator-style plugin modules.
The base class for `Guest` is now `Delegator` instead of `Annotator`.