Show error when refreshing access token fails
When the client is embedded in a page that includes a services
configuration setting then it logs in to Hypothesis using OAuth,
getting an access token from the Hypothesis API.

The Hypothesis API's OAuth access tokens currently expire after an hour
and must be refreshed by the client before then. When the request to
refresh an access token fails the client continues to look and work
exactly as before from the user's point of view, except that all actions
that require communicating with the API such as creating an annotation
fail and show error messages.

Some actions, such as dismissing the sidebar tutorial, fail with no
error message.

Currently we only try the refresh request once for each access token and
don't retry is the request fails. Even if we did retry the refresh
request can't succeed after the access token has expired - it's always
possible for the client to get into a situation where the access token
is expired and can no longer be refreshed.

The only fix (currently) is for the user to reload the page which will
re-do the initial grant token request to get a new access token.

So when a refresh request fails show an error message to the user
telling them that they need to reload the page.