Add *RequestProvided settings
When a on*Request callback function is provided by the host page (for
example onLogoutRequest) add a corresponding onLogoutRequestProvided
boolean to the settings object.

This is because the on*Request functions don't get passed from the host
page context into the sidebar code, because functions aren't JSON
stringifiable, so the sidebar code has no way of knowing whether or not
a callback function was provided.

We're going to be adding code to the sidebar that does need to know
this, so add on*RequestProvided settings to tell it.