Reorganize login-control-test.js by component
Reorganize the tests according to the components (the different items in
the login menu) instead of by context.

Instead of:

    when a first party user is logged in
      it shows the profile, account settings, help and log out buttons
    when a third party user is logged in
      it only shows the profile button (but disabled) and the help button username
    etc

We now have:

    describe: the profile button
      when the user is logged out
        it does not show
      when a first party user is logged in
        it shows the enabled profile button
      when a third party user is logged in
        it shows the enabled profile button
      etc

    describe: the log out button
      etc
    etc

This will make the tests easier to write as we're going to be adding
more complex behavior to some of the buttons, both logic about when the
buttons do and don't show and are enabled or disabled, and logic about
what happens when a button is clicked. It'll be easier to manager if all
the tests for each button are grouped together. eorganize tests by
component