Enable partner sites to customize log out button
The current behavior of the log out button in the sidebar is:

* If the user is logged out it doesn't show
* If the user is logged in to a first-party account it does show
* If the user is logged in to a third-party account it **does not** show

Change the log out button to show when the user is logged in to a
third-party account **and the host page has provided an onLogoutRequest
callback function**. When clicked the log out button will call this
callback function, instead of doing the normal (first-party) log out
procedure.

If the user is logged in to a third-party account and no
`onLogoutRequest` callback function has been provided, then the log out
button will not be shown (as before).