Enable partner sites to customize profile link
Enable partner sites to customize the behavior of the user profile link
in the sidebar's login menu.

The current behavior of this link is:

* If the user is logged out it doesn't show
* If the user is logged in to a first-party account it does show, and
  links to their https://hypothes.is/users/<username> page
* If the user is logged in to a third-party account then it does show,
  and displays their username, but it is just an unclickable `<span>`
  not a link

This commit changes it so that if the user is logged in to a third-party
account and the host page has provided an `onProfileRequest` callback
function, then the profile link will be clickable and clicking it will
call this callback function instead of linking to a hypothes.is user
page.

If the user is logged in to a third-party account but no
`onProfileRequest` function has been provided, then the link will be
shown unclickable as before.